leodos_spacepacket::builder

Struct SpacePacketBuilder

pub struct SpacePacketBuilder<A, B, C, D, E, F> { /* private fields */ }

A typed builder for defining the specification of a SpacePacket.

Implementations

impl<A, B, C, D, E, F> SpacePacketBuilder<A, B, C, D, E, F>

pub fn secondary_header(self, flag: SecondaryHeaderFlag) -> Self

pub fn sequence_flag(self, flag: SequenceFlag) -> Self

impl<B, C, D, E, F> SpacePacketBuilder<Vacant, B, C, D, E, F>

pub fn apid(self, apid: Apid) -> SpacePacketBuilder<Apid, B, C, D, E, F>

impl<A, C, D, E, F> SpacePacketBuilder<A, Vacant, C, D, E, F>

pub fn packet_type( self, packet_type: PacketType, ) -> SpacePacketBuilder<A, PacketType, C, D, E, F>

impl<A, B, D, E, F> SpacePacketBuilder<A, B, Vacant, D, E, F>

pub fn sequence_count( self, count: PacketSequenceCount, ) -> SpacePacketBuilder<A, B, PacketSequenceCount, D, E, F>

impl<A, B, C, E, F> SpacePacketBuilder<A, B, C, Vacant, E, F>

pub fn buffer<'a>( self, buffer: &'a mut [u8], ) -> SpacePacketBuilder<A, B, C, &'a mut [u8], E, F>

impl<A, B, C, D, F> SpacePacketBuilder<A, B, C, D, Vacant, F>

pub fn data_len( self, data_field_len: u16, ) -> SpacePacketBuilder<A, B, C, D, u16, F>

Specifies the length of an uninitialized data field.

pub fn data_type<T: SpacePacketData>( self, ) -> SpacePacketBuilder<A, B, C, D, u16, F>

Specifies the length of an uninitialized data field.

impl<'a> SpacePacketBuilder<Apid, PacketType, PacketSequenceCount, &'a mut [u8], u16, Vacant>

pub fn build(self) -> Result<&'a mut SpacePacket, BuilderError>

Consumes the writer and builds the final SpacePacket view.

impl<'a, A, B, C, F> SpacePacketBuilder<A, B, C, &'a mut [u8], u16, F>

pub fn crc<'b>( self, crc_alg: &'b Crc<u16>, ) -> SpacePacketBuilder<A, B, C, &'a mut [u8], u16, &'b Crc<u16>>

Prepares the builder to construct a packet with a managed CRC.

impl<'a, 'b> SpacePacketBuilder<Apid, PacketType, PacketSequenceCount, &'a mut [u8], u16, &'b Crc<u16>>

pub fn build(self) -> Result<CrcSpacePacket<'a, 'b>, CrcError>