leodos_spacepacket::cfe::tc

Struct Telecommand

#[repr(C)]
pub struct Telecommand<P: SpacePacketData> {
    pub primary: PrimaryHeader,
    pub secondary: TelecommandSecondaryHeader,
    pub payload: P,
}

A zero-copy view over a complete CFE command packet (headers + payload). This is the primary struct you will use to represent a command.

Fields
primary: PrimaryHeadersecondary: TelecommandSecondaryHeaderpayload: P
Implementations

impl Telecommand<()>

pub fn builder() -> TelecommandBuilder<Vacant, Vacant, Vacant, Vacant>

impl<P: SpacePacketData> Telecommand<P>

pub fn new<'a>( buffer: &'a mut [u8], apid: Apid, sequence_count: PacketSequenceCount, function_code: u8, payload: &P, ) -> Result<&'a mut Telecommand<P>, TelecommandError>

pub fn set_cfe_checksum(&mut self)

Calculates and sets the 8-bit cFE checksum for this command packet.

The algorithm is a byte-wise XOR sum of the entire packet, with the checksum field itself treated as zero during calculation.

pub fn validate_cfe_checksum(&self) -> bool

Validates the 8-bit cFE checksum.

Returns true if the checksum is valid, false otherwise.