leodos_spacepacket::tctm::tc::builder

Struct TCFrameBuilder

pub struct TCFrameBuilder<A, B, C, D> { /* private fields */ }

A typestate builder for TCTransferFrame.

Implementations

impl<A, B, C, D> TCFrameBuilder<A, B, C, D>

pub fn bypass_flag(self, flag: BypassFlag) -> Self

Sets the Bypass Flag (optional). Defaults to TypeA.

pub fn control_flag(self, flag: ControlFlag) -> Self

Sets the Control Flag (optional). Defaults to TypeD.

pub fn sequence_num(self, seq: u8) -> Self

Sets the Frame Sequence Number (optional). Defaults to 0.

impl<B, C, D> TCFrameBuilder<Vacant, B, C, D>

pub fn scid(self, scid: u16) -> TCFrameBuilder<u16, B, C, D>

Sets the Spacecraft ID (SCID), a required field.

impl<A, C, D> TCFrameBuilder<A, Vacant, C, D>

pub fn vcid(self, vcid: u8) -> TCFrameBuilder<A, u8, C, D>

Sets the Virtual Channel ID (VCID), a required field.

impl<A, B, D> TCFrameBuilder<A, B, Vacant, D>

pub fn buffer<'a>( self, buffer: &'a mut [u8], ) -> TCFrameBuilder<A, B, &'a mut [u8], D>

Sets the memory buffer where the frame will be written, a required field.

impl<A, B, C> TCFrameBuilder<A, B, C, Vacant>

pub fn data_field_len( self, data_field_len: usize, ) -> TCFrameBuilder<A, B, C, usize>

Sets the length of the frame’s data field in bytes, a required field.

impl<'a> TCFrameBuilder<u16, u8, &'a mut [u8], usize>

pub fn build(self) -> Result<&'a mut TCTransferFrame, BuildError>

Consumes the builder and constructs the TCTransferFrame header.

This method is only available after all required fields (scid, vcid, buffer, and data_field_len) have been set.

Returns a mutable view of the frame, ready for the user to write data into its data field.