leodos_spacepacket::tctm::tc

Struct TCTransferFrame

#[repr(C, packed(1))]
pub struct TCTransferFrame { /* private fields */ }

A zero-copy view over a TC Transfer Frame in a raw byte buffer.

This struct can be created via TCTransferFrame::builder() and provides access to the frame header and its data field, which typically contains one or more SpacePackets.

Implementations

impl TCTransferFrame

pub const HEADER_SIZE: usize = 5usize

The size of the TC Transfer Frame header in bytes.

pub const MAX_DATA_FIELD_LEN: usize = 1_019usize

The maximum allowed size of the data field in bytes.

pub fn builder() -> TCFrameBuilder<Vacant, Vacant, Vacant, Vacant>

Creates a new builder to begin constructing a TCTransferFrame.

pub fn header(&self) -> &TCHeader

Returns a reference to the frame’s header.

pub fn data_field_mut(&mut self) -> &mut [u8]

Returns a mutable reference to the frame’s data field.

This is typically used to copy a serialized SpacePacket into the frame.

pub fn frame_len(&self) -> usize

Returns the total length of the frame (header + data field) in bytes.