leodos_spacepacket::tctm::tm

Struct TMTransferFrame

#[repr(C, packed(1))]
pub struct TMTransferFrame { /* private fields */ }

A zero-copy view over a TM Transfer Frame in a raw byte buffer.

Unlike TC frames, TM frames are typically only parsed, not built, by ground software. They have a fixed length per physical channel, which must be known in advance.

Implementations

impl TMTransferFrame

pub fn parse<'a>( bytes: &[u8], output_buffer: &'a mut [u8], randomizer: &impl Randomizer, ) -> Result<&'a TMTransferFrame, ParseError>

Parses a raw, possibly randomized, byte slice into a zero-copy TM Transfer Frame view.

The incoming bytes slice is de-randomized into the output_buffer. The The returned &TMTransferFrame is a view over this output_buffer.

Arguments

bytes: The raw bytes of the TM frame received from the radio.

output_buffer: A buffer at least as large as bytes to hold the de-randomized frame.

randomizer: The specific randomization algorithm to apply.

pub fn header(&self) -> &TMHeader

Returns a reference to the frame’s header.

pub fn data_field(&self) -> &[u8]

Returns a reference to the frame’s data field.

This slice typically contains one or more SpacePackets that can now be parsed individually.